+++ WEB/FE/src/components/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route, Link } from 'react-router-dom';
 import * as Pages from '@pages/index';
-import { PrivateRoute } from '@components/route/PrivateRoute';
+import { PrivateRoute } from '@/components/PrivateRoute/PrivateRoute';
 import ComfirmEmail from '@components/confirmEmail/index';
 import { Modal } from './common/Modal';
